pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "panoseti-grpc"
version = "0.2.5"
description = "gRPC for the PANOSETI project."
readme = "README.md"
requires-python = ">=3.9"
authors = [
    { name="Nicolas Rault-Wang"},
    { name="Ben Godfrey"}
]
license-files = [ "LICENSE" ]
keywords = [
    "panoseti", "gRPC", "observatory", "data-acquisition", "astronomy", "real-time"
]
classifiers = [
    "Programming Language :: Python :: 3",
    "Operating System :: Unix",
    "Intended Audience :: Science/Research",
    "Topic :: Scientific/Engineering :: Astronomy",
    "Framework :: AsyncIO",
]

dependencies = [
    "grpcio>=1.76.0",
    "grpcio-reflection>=1.76.0",
    "grpcio-tools>=1.76.0",
    "protobuf>=6.33.4,<7.0.0",
    "coverage>=7.10.7",
    "cython",
    "wheel",
    "rich",
    "redis",
    "snakeviz",
    "json5",
    "pyserial",
    "pyubx2",
    "seaborn",
    "matplotlib",
    "pandas",
    "numpy",
    "psutil",
    "watchfiles",
    "aiofiles",
    "pygnssutils",
    "pydantic",
    "tomli"
]

[project.optional-dependencies]
dev = [
    "pytest",
    "black",
    "flake8",
    "ipython",
]
docs = [
    "sphinx",
    "myst-parser",
    "sphinx-autodoc-typehints",
]

[project.urls]
Homepage = "https://github.com/panoseti/panoseti_grpc"
Documentation = "https://github.com/panoseti/panoseti_grpc/blob/main/README.md"
Repository = "https://github.com/panoseti/panoseti_grpc"
Issues = "https://github.com/panoseti/panoseti_grpc/issues"

[tool.setuptools.packages.find]
where = ["src"]
#include = [
#    "panoseti_grpc/daq_data/*",
#    "ublox_control/*",
#    "generated/*",
#    "panoseti_util/*",
#]
exclude = ["*logs*", "*.pcapng", "*.png"]

# Include non-python data files:
[tool.setuptools.package-data]
"*" = ["*.json", "*.json5", "*.txt", "*.csv", "*.pff", "*.pffd", "*.toml"]
"panoseti_grpc.daq_data" = ["simulated_data_dir/**/*"]

[tool.setuptools]
include-package-data = true

[tool.black]
line-length = 120
target-version = ["py39"]

[tool.flake8]
max-line-length = 120
extend-ignore = ["E203", "W503"]

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-ra -q"
testpaths = [
    "tests",
]

